Do not push command if mouse is over bottom panel
