+++ Zilon.Core/CDT.LAST.MonoGameClient/ViewModels/MainScene/Ui/CombatActPanel.cs
@@ -47,19 +47,6 @@ namespace CDT.LAST.MonoGameClient.ViewModels.MainScene.Ui
             _equipmentModule.EquipmentChanged -= EquipmentModule_EquipmentChanged;
         }
 
-        private void DrawBackground(SpriteBatch spriteBatch, GraphicsDevice graphicsDevice)
-        {
-            const int PANEL_MARGIN = 4;
-            const int PANEL_WIDTH = (32 * 8) + 16 + PANEL_MARGIN;
-            const int PANEL_HEIGHT = 32 + (4 * 2);
-
-            var panelX = (graphicsDevice.Viewport.Width - PANEL_WIDTH) / 2;
-
-            spriteBatch.Draw(_uiContentStorage.GetBottomPanelBackground(),
-                new Rectangle(panelX, graphicsDevice.Viewport.Height - PANEL_HEIGHT, PANEL_WIDTH, PANEL_HEIGHT),
-                Color.White);
-        }
-
         private void DrawButtonHotkey(int actIndex, ButtonBase button, SpriteBatch spriteBatch)
         {
             var spriteFont = _uiContentStorage.GetAuxTextFont();
@@ -145,8 +132,6 @@ namespace CDT.LAST.MonoGameClient.ViewModels.MainScene.Ui
 
         public void Draw(SpriteBatch spriteBatch, GraphicsDevice graphicsDevice)
         {
-            DrawBackground(spriteBatch, graphicsDevice);
-
             const int COMBAT_ACT_BUTTON_SIZE = 32;
             const int BOTTOM_MARGIN = 0;
