+++ Zilon.Core/CDT.LAST.MonoGameClient/ViewModels/MainScene/Ui/TravelMenuPanel.cs
@@ -154,14 +154,6 @@ namespace CDT.LAST.MonoGameClient.ViewModels.MainScene.Ui
             var bottomOfScreenY = graphicsDevice.Viewport.Height;
 
             const int PANEL_MARGIN = 4;
-            const int PANEL_WIDTH = (32 * 8) + 16 + PANEL_MARGIN;
-            const int PANEL_HEIGHT = 32 + (4 * 2);
-
-            var panelX = (graphicsDevice.Viewport.Width - PANEL_WIDTH) / 2;
-
-            spriteBatch.Draw(_uiContentStorage.GetBottomPanelBackground(),
-                new Rectangle(panelX, graphicsDevice.Viewport.Height - PANEL_HEIGHT, PANEL_WIDTH, PANEL_HEIGHT),
-                Color.White);
 
             for (var buttonIndex = 0; buttonIndex < _buttons.Length; buttonIndex++)
             {
